+++ src/features/atlas/settings/components/MapSettingsPanel.tsx
@@ -6,6 +6,7 @@ import { useUI } from "@contexts/UIContext";
 import { useAccessibility } from "@features/settings";
 import { ColorsSettingsGroup } from "./ColorsSettingsGroup";
 import { ConfigurationSettingsGroup } from "./ConfigurationSettingsGroup";
+import { InterfaceSettingsGroup } from "./InterfaceSettingsGroup";
 import { OverlaySettingsGroup } from "./OverlaySettingsGroup";
 
 export function MapSettingsPanel() {
@@ -28,6 +29,8 @@ export function MapSettingsPanel() {
     >
       <div className="mt-4">
         <ConfigurationSettingsGroup />
+        <Separator className="my-4" />
+        <InterfaceSettingsGroup />
         <Separator className="my-4" />
         <OverlaySettingsGroup />
         <Separator className="my-4" />