+++ src/features/atlas/ui/layouts/MapUiContainer.tsx
@@ -12,8 +12,9 @@ import {
   useMapLegendItems,
   type LegendItem,
 } from "@features/atlas/legend";
+import { useMapInterfaceSettings } from "@features/atlas/settings";
 import { TimelineBar, TimelineNavigator } from "@features/atlas/timeline";
-import { useUiHint } from "@hooks";
+import { useScreenSize, useUiHint } from "@hooks";
 import { MapToolbar } from "../components/controls/MapToolbar";
 import { MapFooter } from "../components/footer/MapFooter";
 
@@ -27,6 +28,7 @@ export function MapUiContainer({
   isEmbed,
 }: MapUiContainerProps) {
   const effectiveLayers = useEffectiveLayers();
+  const { toolbarOrientation } = useMapInterfaceSettings();
   const {
     isReadonly,
     isEdit,
@@ -37,15 +39,16 @@ export function MapUiContainer({
     center,
     selectedCoords,
   } = useMapView();
+  const { isLaptop } = useScreenSize();
   const { timelineMode } = useTimeline();
   const { uiVisible, openUserPanel, showLegend, closeLegend } = useUI();
+  const { t } = useTranslation("atlas");
+
   const legendItems: LegendItem[] = useMapLegendItems(
     effectiveLayers,
     timelineMode,
     colorMode,
   );
-
-  const { t } = useTranslation("atlas");
 
   // UI hint for adding marker
   const addMarkerHint = useMemo(
@@ -119,6 +122,9 @@ export function MapUiContainer({
   // Don't render UI if not visible
   if (!uiVisible) return null;
 
+  const effectiveToolbarOrientation =
+    isLaptop && timelineMode ? "vertical" : toolbarOrientation;
+
   return (
     <>
       {timelineMode && !isEmbed && !isEdit && (
@@ -129,7 +135,12 @@ export function MapUiContainer({
       )}
       {!openUserPanel && (
         <>
-          <MapToolbar zoom={zoom} setZoom={setZoom} isEmbed={isEmbed} />
+          <MapToolbar
+            orientation={effectiveToolbarOrientation}
+            zoom={zoom}
+            setZoom={setZoom}
+            isEmbed={isEmbed}
+          />
           {!isEmbed && (
             <>
               <MapFooter